Review fix: share underflow JUL capture helpers in BTreeEngineTestFixtures
Promote readAtomicLong, readAtomicLongRef, installCapturingHandler, and
the new captureSevereOn lambda runner into the existing fixtures class so
the single-value and multi-value underflow suites stop carrying three
byte-identical helpers each. The six straight-line tests in each suite
(those whose engine invocations all land before the assertion phase)
collapse onto captureSevereOn; the two tests that interleave assertions
with engine calls inside the capture window keep their manual handler
ritual and now reach the shared helper by qualified call.

Removes ~200 lines per file with no change in test count or contract.
Aligns with the Step 6 storage cascade-containment test that would
otherwise have introduced a third copy of the same scaffolding.

YTDB-958
